Funcionalidad a controladores y se realizo una migracion a la base de datos

+++ Estacionamiento/Entidades/Estacionamientos.cs
@@ -5,6 +5,6 @@
         public int id { get;set; }
         public string Dir { get;set; }  
         public int EspMax { get;set; }
-        public int Tarifa { get;set; }
+        public float Tarifa { get;set; }
     }
 }
